Create signin-form component, render it with App
Moved ajax request out of App.js and into signin-form.js
Next step is to lift data out of signin-form and into app, so that
other components can access it. For instance, a token is required for
all CRUD actions, so the token that is created from logging in needs
to be saved up in the top layer App component.