Add signUp form and catalog message to new users
Users can now create new accounts.
Changed Guitars component to show a message if users don't have any
guitars, because guitars.map was breaking the site if they didn't have
any.
I was trying to do if this.state.guitars.length === 0, but the problem
with that is that I overwrite this.state.guitars with whatever the AJAX
request returns (on line 27). Therefore it was 'undefined'.