Commit setups before refactoring AJAX requests
Commented out buggy setups code because I'm going to try refactoring
the whole app so all AJAX functions exist in one file rather than a
bunch in the App component and others in separate components.

In conjunction with the store that can save user token and even guitar
and setup ids, this should be a worthwhile refactoring.