Refactor UpdateGuitar component
Fixed an issue where React wasn't happy that the component was both
controlled and uncontrolled. The answer was letting the forms control
the state (as they had always done) but instead of setting state in the
constructor function to be props, setting it to empty strings, and
using componentDidMount to get the initial data (from the props).

Also designed the update form to hide and show on a button click

Also brought in the API call from api.js