Fix CreateSetup component
CreateSetup component had been mostly copy/pasted from CreateGuitar
so there was a lot of bad information in it. It wasn't being used in
the program yet so it got saved like that for a few commits.

Now it is being used by the Setups component and has been fixed up.
TODO: add .then and .catch after creation of Setup