Filter setups and save first setup in state
Before this commit, every Setup component (one for every guitar) was
saving every guitars setups in their state. I wasn't thinking about them
as multiple components that could each have their own version of state.

Then in the render method, setups were filtered so that only setups
belonging to the specific guitar were rendered on the page. newestSetup
was then chosen as the first of these.

Now the filtering is done and then the filtered results are set in state
so state only ever saves setups that belong to the specific guitar.
Not only does this seem like better practice, but actually saving
newestSetup in state was necessary to refresh the view every time a
setup was created.

Before, const newestSetup was set to the first setup and then was static
It wouldn't update when a newer newestSetup was created. By saving
it in state, the component now knows to refresh the view whenever a
new setup is created.